refactor: added exception to createMeal

+++ src/storage/meals/createMeal/index.tsx
@@ -7,6 +7,7 @@ import { MealDTO } from "../MealDTO";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import "react-native-get-random-values";
 import { v4 as uuidv4 } from "uuid";
+import { AppException } from "@utils/AppException";
 
 export const createMeal = async ({
   name,
@@ -26,6 +27,14 @@ export const createMeal = async ({
     };
 
     const mealsStoraged = await getAllMeals();
+    const mealAlreadyExists = mealsStoraged.includes(newMeal);
+
+    if (mealAlreadyExists) {
+      throw new AppException(
+        "Essa refeição já está cadastrada em nosso banco de dados!"
+      );
+    }
+
     const mealsUpdated = [...mealsStoraged, newMeal];
 
     await AsyncStorage.setItem(MEALS_COLLECTION, JSON.stringify(mealsUpdated));
